Fix trace for segments of varying lengths
Planar's `Ray` automatically scales the direction of the ray to unit
length, but `line_segment` assumed that the length of the segment is
preserved in the corresponding `Ray`'s `direction`. The current change
keeps track of the segment's length separately to determine whether the
intersection point falls within the segment. See issue #91 on GitHub:

https://github.com/danoneata/chalk/issues/91