TYPE Move required arguments as object attributes
In order to make the code pass the typing checks, I've moved the
arguments that were required when visiting the AST as attributes of the
instanec (see for example `dwg` in `backend/svg.py`). I was able to move
only the immutable arguments, while the ones that change during the
traversal of hierarchy had to be kept as method arguments (otherwise we
run into hairy bugs).